Formatting mostly.... aaaaannnnd confusion!
Umm... ok.. did anyone actually read what we have to do? Because I don't
understand at all.. He says we need 5 documentation, 10 implementations,
and 10 'how to take advantage of the vulnerability'-s?? Is this for each
vulnerability? Even if it isn't, then why do we have to do 5/10 things
when we only have 3 vulnerabilities????
I'm so confused! >__<